requirements.txt
discord-py
ffmpeg-python
aiohttp
aiofiles
requests
beautifulsoup4
python-dotenv